# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: github.com/is7qin/c3api Admin API
  version: 1.0.0
  description: '管理端 API（模板/账号/分组/规则/日志/统计）。认证：Authorization: Bearer <admin_token>。'
servers:
  - url: /admin
...
paths:
...
  /groups:
    post:
      operationId: PostGroups
      tags: [groups]
      summary: 创建分组（平台容量池；key 为独立表，由用户面 /user/keys 创建）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupCreate' }
      responses:
        '200':
          description: 创建后的分组
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        default:
          $ref: '#/components/responses/Error'
    get:
      operationId: GetGroups
      tags: [groups]
      summary: 分组列表（分页/筛选/排序）
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 200, description: 每页条数（上限 200，超限裁剪到 200） } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: name, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200':
          description: 分组列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GroupListResponse' }
        default:
          $ref: '#/components/responses/Error'
  /groups/batch-delete:
    post:
      operationId: PostGroupsBatchDelete
      tags: [groups]
      summary: 批量删除分组（事务，全成或全败）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchDeleteBody' }
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchDeleteResponse' }
        default:
          $ref: '#/components/responses/Error'
  /groups/batch-update:
    post:
      operationId: PostGroupsBatchUpdate
      tags: [groups]
      summary: 批量更新分组（fields 为任意字段子集）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchUpdateGroupsBody' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchUpdateResponse' }
        default:
          $ref: '#/components/responses/Error'
  /groups/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetGroupsId
      tags: [groups]
      responses:
        '200':
          description: 分组
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: PutGroupsId
      tags: [groups]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupCreate' }
      responses:
        '200':
          description: 更新后分组
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        default:
          $ref: '#/components/responses/Error'
    delete:
      operationId: DeleteGroupsId
      tags: [groups]
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletedResponse' }
        default:
          $ref: '#/components/responses/Error'
...
  /groups/{id}/assignments:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetGroupsIdAssignments
      tags: [groups]
      summary: 读取组的授予用户与专属倍率（platform_admin；与 PUT 对称，供前端预填充与安全全量写回）
      responses:
        '200':
          description: 该组当前的授予 user_id 列表与专属倍率
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GroupAssignmentsResponse' }
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: PutGroupsIdAssignments
      tags: [groups]
      summary: 设置组的授予用户（platform_admin；替换语义：未列出即撤销，空数组 = 清空）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GroupAssignmentsBody' }
      responses:
        '200':
          description: 已生效的授予 user_id 列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GroupAssignmentsResponse' }
        default:
          $ref: '#/components/responses/Error'
...
  /redemption-codes:
    post:
      operationId: PostRedemptionCodes
      tags: [redemptions]
      summary: 生成兑换码（1..count 个；count 默认 1，上限 1000；type 枚举/value>0/temp_balance 必填 resource_expires_at）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateRequest' }
      responses:
        '200':
          description: 生成的完整码列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerateResponse' }
        default:
          $ref: '#/components/responses/Error'
    get:
      operationId: GetRedemptionCodes
      tags: [redemptions]
      summary: 兑换码列表（增强分页范式 page/page_size；type/status 筛选；sort 白名单）
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1, minimum: 1, description: 页码（1-based） } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, minimum: 1, maximum: 1000 } }
        - { name: type, in: query, schema: { type: string, enum: [balance, concurrency, temp_balance] } }
        - { name: status, in: query, schema: { type: string, enum: [active, disabled] } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200':
          description: 兑换码列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RedemptionCodeListResponse' }
        default:
          $ref: '#/components/responses/Error'
  /redemption-codes/batch-deactivate:
    post:
      operationId: PostRedemptionCodesBatchDeactivate
      tags: [redemptions]
      summary: 批量失效兑换码（单事务；已失效 no-op；缺失 id → 404 含缺失详情）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchDeactivateRequest' }
      responses:
        '200':
          description: 新失效数（已 disabled 不计）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchDeactivateResponse' }
        default:
          $ref: '#/components/responses/Error'
  /redemption-codes/{id}/uses:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetRedemptionCodesIdUses
      tags: [redemptions]
      summary: 某码的兑换记录（审计；码缺失 → 404）
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: 兑换记录列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RedemptionUseListResponse' }
        default:
          $ref: '#/components/responses/Error'
  /redemption-codes/{id}/deactivate:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      operationId: PostRedemptionCodesIdDeactivate
      tags: [redemptions]
      summary: 单码失效（已失效 no-op；缺失 → 404）
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeactivateRequest' }
      responses:
        '200':
          description: 失效成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeactivateResponse' }
        default:
          $ref: '#/components/responses/Error'
  # --- litellm 模型价格（tag: pricing；Phase 5 计费价格来源） ---
  # 模型标识在 query（PUT/DELETE 的 model 参数）：模型名是自由字符串可含 `/`
  # （如 1024-x-1024/50-steps/bedrock/amazon.nova-canvas-v1:0），路径参数单段
  # 匹配会拆段 404——URL 编码与 chi 通配方案均否决，见 spec 裁决（2026-08-14）。
...
components:
  responses:
    Error:
      description: 错误响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    DeletedResponse:
      type: object
      required: [deleted]
      properties:
        deleted: { type: boolean }
...
    GroupListResponse:
      type: object
      required: [total, rows]
      properties:
        total: { type: integer, format: int64 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/Group' }
    GroupVisibility:
      type: string
      enum: [public, private]
    GroupProtocolConvert:
      type: string
      enum: [chat_to_resp, mess_to_resp, resp_to_mess, chat_to_mess]
      description: 分组级协议转换方向（只补差，W5 网关 internal/protoconv 消费）：chat_to_resp = 客户端 chat → 模板 resp；mess_to_resp = anthropic messages → resp；resp_to_mess = resp → anthropic messages；chat_to_mess = chat → anthropic messages。off 不在枚举内——不转换 = 空数组（见 protocol_convert 字段说明）
    GroupAssignmentsBody:
      type: object
      required: [user_ids]
      properties:
        user_ids:
          type: array
          items: { type: integer, format: int64 }
          description: 替换语义：完整授予列表（未列出即撤销；空数组 = 清空）
        multipliers:
          type: object
          additionalProperties:
            type: number
            format: double
            nullable: true
            minimum: 0
            maximum: 10
          description: 可选：user_id → 该用户在该组的专属价格倍率（正常值，1 = ×1，0 = 免费，上限 10 = ×10；API 边界与万分数换算）。仅对 user_ids 中列出的用户生效；null = 清除为未设置（回退组倍率）；未列出的用户沿用当前值
    GroupAssignmentsResponse:
      type: object
      required: [user_ids]
      properties:
        user_ids:
          type: array
          items: { type: integer, format: int64 }
        multipliers:
          type: object
          additionalProperties:
            type: number
            format: double
            nullable: true
          description: 该组当前各用户的专属价格倍率（正常值；null/缺省 = 未设置 → 用组倍率）
...
    GroupCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        visibility: { $ref: '#/components/schemas/GroupVisibility' }
        price_multiplier: { type: number, format: double, minimum: 0, maximum: 10, nullable: true, description: 价格倍率（正常值，1 = ×1，0 = 免费，上限 10 = ×10；API 边界与万分数换算——存储 15000 ↔ 显示 1.5）。缺省/null = 不设置（×1）；显式 0 = 免费组；PUT 显式写（含 0） }
        protocol_convert:
          type: array
          items: { $ref: '#/components/schemas/GroupProtocolConvert' }
          description: 协议转换方向集合（多方向并存，按客户端格式命中——chat 请求走 chat_to_*、anthropic 请求走 mess_to_resp、resp 请求走 resp_to_mess）。空数组/缺省 = off = 不转换；同客户端格式多方向（chat_to_resp 与 chat_to_mess 并存）→ 400。PUT 语义：缺省（null/省略）= 保持原值；显式空数组 = 清空既有方向
...
    Group:
      type: object
      properties:
        ID: { type: integer, format: int64 }
        Name: { type: string }
        Visibility: { $ref: '#/components/schemas/GroupVisibility' }
        PriceMultiplier: { type: number, format: double, description: 价格倍率（正常值，1 = ×1，0 = 免费，上限 10 = ×10；API 边界与万分数换算——存储 15000 ↔ 显示 1.5） }
        ProtocolConvert:
          type: array
          items: { $ref: '#/components/schemas/GroupProtocolConvert' }
          description: 协议转换方向集合（空数组 = off = 不转换；多方向按客户端格式命中）
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
        DeletedAt: { type: string, format: date-time, nullable: true, description: 软删除时间戳；null = 存活（列表/消费路径过滤已删；GET 单个可查已删项） }
...
    BatchDeleteBody:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items: { type: integer, format: int64 }
          minItems: 1
          maxItems: 100
...
    GroupPatch:
      type: object
      properties:
        name: { type: string }
        visibility: { $ref: '#/components/schemas/GroupVisibility' }
...
    BatchUpdateGroupsBody:
      type: object
      required: [ids, fields]
      properties:
        ids:
          type: array
          items: { type: integer, format: int64 }
          minItems: 1
          maxItems: 100
        fields: { $ref: '#/components/schemas/GroupPatch' }
    BatchDeleteResponse:
      type: object
      required: [deleted]
      properties:
        deleted: { type: integer }
    BatchUpdateResponse:
      type: object
      required: [updated]
      properties:
        updated: { type: integer }
    # --- 兑换码（Phase 5 计费前基础设施：资源发放的通用载体） ---
    RedemptionType:
      type: string
      enum: [balance, concurrency, temp_balance]
    RedemptionStatus:
      type: string
      enum: [active, disabled]
    RedemptionCode:
      type: object
      required: [ID, Code, Type, Value, MaxUses, UsedCount, Status, CreatedBy, CreatedAt, UpdatedAt]
      properties:
        ID: { type: integer, format: int64 }
        Code: { type: string, description: XXXX-XXXX-XXXX-XXXX（16 字符，生成后不可编辑，仅可失效） }
        Type: { $ref: '#/components/schemas/RedemptionType' }
        Value: { type: number, format: double, description: 面值：balance/temp_balance = USD（1 = $1）；concurrency = 并发数（整数）。存储毫分（API 边界换算） }
        Remark: { type: string, nullable: true }
        ExpiresAt: { type: string, format: date-time, nullable: true, description: 码未兑换即过期；null = 永久 }
        ResourceExpiresAt: { type: string, format: date-time, nullable: true, description: 兑换后资源到期；temp_balance 必有，其余恒 null }
        MaxUses: { type: integer, description: 1 = 单次码；>1 = 多人码 }
        UsedCount: { type: integer, description: 已兑换次数 }
        Status: { $ref: '#/components/schemas/RedemptionStatus' }
        CreatedBy: { type: integer, format: int64, description: 0 = 系统（静态 admin token）；>0 = platform_admin 用户 id }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
    RedemptionUse:
      type: object
      required: [ID, CodeID, UserID, Value, CreatedAt]
      properties:
        ID: { type: integer, format: int64 }
        CodeID: { type: integer, format: int64 }
        UserID: { type: integer, format: int64 }
        Value: { type: number, format: double, description: 兑换时的值快照（同面值单位语义） }
        ResourceExpiresAt: { type: string, format: date-time, nullable: true }
        CreatedAt: { type: string, format: date-time }
...
    RedemptionCodeListResponse:
      type: object
      required: [total, rows]
      properties:
        total: { type: integer, format: int64 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/RedemptionCode' }
    RedemptionUseListResponse:
      type: object
      required: [total, rows]
      properties:
        total: { type: integer, format: int64 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/RedemptionUse' }
...
    GenerateRequest:
      type: object
      required: [type, value]
      properties:
        type: { $ref: '#/components/schemas/RedemptionType' }
        value: { type: number, format: double, description: 面值：balance/temp_balance = USD；concurrency = 并发数（整数）。> 0 }
        remark: { type: string }
        expires_at: { type: string, format: date-time, description: 码未兑换即过期；缺省 = 永久 }
        resource_expires_at: { type: string, format: date-time, description: 兑换后资源到期；temp_balance 必填（决策 4） }
        max_uses: { type: integer, description: 1 = 单次码（缺省）；>1 = 多人码 }
        count: { type: integer, description: 生成个数 1-1000（缺省 1） }
    GenerateResponse:
      type: object
      required: [codes]
      properties:
        codes:
          type: array
          items: { $ref: '#/components/schemas/RedemptionCode' }
    DeactivateRequest:
      type: object
      description: 单码失效无请求体（保留空 schema 对齐批量端点命名）
    DeactivateResponse:
      type: object
      required: [deactivated]
      properties:
        deactivated: { type: boolean }
    BatchDeactivateRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items: { type: integer, format: int64 }
          minItems: 1
          maxItems: 100
    BatchDeactivateResponse:
      type: object
      required: [deactivated]
      properties:
        deactivated: { type: integer, description: 新失效数（已 disabled no-op 不计） }
    # --- litellm 模型价格（Phase 5 计费价格来源） ---
...